refactor(mm-next): move server-side only utils to specific directory
