+++ packages/mirror-media-next/pages/recover-password/index.js
@@ -10,7 +10,7 @@ import PrimaryButton from '../../components/shared/buttons/primary-button'
 import TextButton from '../../components/login/text-button'
 import { getLogTraceObject, isValidEmail } from '../../utils'
 import { handleAxiosResponse } from '../../utils/response-handle'
-import redirectToDestinationWhileAuthed from '../../utils/redirect-to-destination-while-authed'
+import redirectToDestinationWhileAuthed from '../../utils/server-side-only/redirect-to-destination-while-authed'
 import { auth } from '../../firebase'
 import { FirebaseError } from 'firebase/app'
 import { sendPasswordResetEmail } from 'firebase/auth'
