+++ packages/mirror-media-next/pages/update-password/index.js
@@ -4,7 +4,7 @@ import { getLogTraceObject } from '../../utils'
 import { handleAxiosResponse } from '../../utils/response-handle'
 import { fetchHeaderDataInDefaultPageLayout } from '../../utils/api'
 import { getSectionAndTopicFromDefaultHeaderData } from '../../utils/data-process'
-import withUserSSR from '../../utils/with-user-ssr'
+import withUserSSR from '../../utils/server-side-only/with-user-ssr'
 import { getLoginUrl } from '../../utils/server-side-only'
 import useMembershipRequired from '../../hooks/use-membership-required'
 import LayoutFull from '../../components/shared/layout-full'
