+++ packages/mirror-media-next/utils/server-side-only/redirect-to-destination-while-authed.js
