+++ packages/mirror-media-next/utils/server-side-only/redirect-to-login-while-unauthed.js
@@ -1,5 +1,5 @@
 import withUserSSR from './with-user-ssr'
-import { getLoginUrl } from './server-side-only'
+import { getLoginUrl } from './index'
 
 /**
  * @typedef {import('querystring').ParsedUrlQuery} ParsedUrlQuery
