+++ packages/mirror-media-next/utils/server-side-only/with-user-ssr.js
@@ -1,4 +1,4 @@
-import { getAdminAuth } from '../firebase/admin'
+import { getAdminAuth } from '../../firebase/admin'
 
 /**
  * @typedef {import('querystring').ParsedUrlQuery} ParsedUrlQuery
